+++ src/Phan/Language/EmptyUnionType.php
@@ -1053,6 +1053,16 @@ final class EmptyUnionType extends UnionType
         return false;
     }
 
+    public function isNonNullNumberType() : bool
+    {
+        return false;
+    }
+
+    public function hasStringType() : bool
+    {
+        return false;
+    }
+
     public function hasNonNullStringType() : bool
     {
         return false;